:construction: Moved logic to frontend
